refactor(sdk-utilities): extract handleMicroExit helper from oidc-client
